Feat/#240 inactive

See merge request s12-bigdata-dist-sub1/S12P21A505!200

+++ frontend/src/features/auth/useAuth.ts
@@ -1,7 +1,8 @@
 import { useNavigate } from "react-router";
 import useToast from "@/shared/hooks/useToast";
-import { login } from "@/shared/api/auth";
+import { inactiveAPI, login, logoutAPI } from "@/shared/api/auth";
 import { UserType } from "./type";
+import { useMutation } from "@tanstack/react-query";
 
 export function useAuth() {
   const navigate = useNavigate();
@@ -40,10 +41,16 @@ export function useAuth() {
     return user && token;
   }
 
-  function logout() {
-    sessionStorage.clear();
-    navigate("/login");
-    showToast({ message: "로그아웃되었습니다.", type: "success" });
+  async function logout() {
+    const data = await logoutAPI();
+    const { success, errorMsg } = data;
+    if (success) {
+      sessionStorage.clear();
+      navigate("/login");
+      showToast({ message: "로그아웃되었습니다.", type: "success" });
+    } else {
+      showToast({ message: errorMsg || "로그아웃 실패", type: "error" });
+    }
   }
 
   function getUserMeta(): UserType {
@@ -52,5 +59,36 @@ export function useAuth() {
     return { id, nickname, s3Url };
   }
 
-  return { handleLogin, isLoggedIn, logout, getUserMeta };
+  const mutation = useMutation<boolean, Error, void>({
+    mutationFn: inactiveAPI,
+    onSuccess: (data) => {
+      const { success, errorMsg } = data as any;
+      if (!success) {
+        showToast({
+          message: errorMsg || "회원 탈퇴를 실패했습니다.",
+          type: "error",
+        });
+      } else {
+        sessionStorage.clear();
+        showToast({ message: "회원 탈퇴가 완료되었습니다.", type: "success" });
+      }
+    },
+    onError: () => {
+      showToast({
+        message: "회원 탈퇴 중 오류가 발생했습니다.",
+        type: "error",
+      });
+    },
+  });
+
+  const deleteAccount = () => mutation.mutate();
+
+  return {
+    handleLogin,
+    isLoggedIn,
+    logout,
+    getUserMeta,
+    deleteAccount,
+    isDeleting: mutation.isPending,
+  };
 }
